WorldScreenMusicPopup / SoundTab - move music controls to their own file
